Move prefrence provider in to its folder
